getCurrentMatchupsByUser endpoint and query

+++ src/main/java/com/wcappel/ffbackend/dto/MatchupMetaDTO.java
@@ -7,8 +7,8 @@ public interface MatchupMetaDTO {
     int getLeague_match_num();
     String gethomeTeam();
     String getawayTeam();
-    int getHome_score();
-    int getAway_score();
+    Float getHome_score();
+    Float getAway_score();
     String getHome_logo();
     String getAway_logo();
     String getHome_owner();
